backend/requirements.txt
fastapi
uvicorn
python-multipart
whisperx
deep-translator
edge-tts
pyannote.audio
torchaudio
pandas
pydub
SpeechRecognition
pygame
python-dotenv
